ref(sampling): Remove error rules code
