+++ static/app/views/settings/project/filtersAndSampling/rules/rule/type.tsx
@@ -10,8 +10,6 @@ type Props = {
 
 function Type({type}: Props) {
   switch (type) {
-    case DynamicSamplingRuleType.ERROR:
-      return <ErrorLabel>{t('Errors only')}</ErrorLabel>;
     case DynamicSamplingRuleType.TRANSACTION:
       return <TransactionLabel>{t('Individual transactions')}</TransactionLabel>;
     case DynamicSamplingRuleType.TRACE:
