+++ tests/js/spec/views/settings/projectFiltersAndSampling/transactions.spec.tsx
@@ -14,35 +14,22 @@ import {
 import {commonConditionCategories, renderComponent, renderModal} from './utils';
 
 describe('Filters and Sampling - Transaction rule', function () {
+  beforeEach(() => {
+    MockApiClient.addMockResponse({
+      url: '/projects/org-slug/project-slug/tags/',
+      method: 'GET',
+      body: TestStubs.Tags,
+    });
+  });
+
   describe('transaction rule', function () {
     it('renders', async function () {
-      MockApiClient.addMockResponse({
-        url: '/projects/org-slug/project-slug/tags/',
-        method: 'GET',
-        body: TestStubs.Tags,
-      });
-
       MockApiClient.addMockResponse({
         url: '/projects/org-slug/project-slug/',
         method: 'GET',
         body: TestStubs.Project({
           dynamicSampling: {
             rules: [
-              {
-                sampleRate: 0.1,
-                type: 'error',
-                condition: {
-                  op: 'and',
-                  inner: [
-                    {
-                      op: 'glob',
-                      name: 'event.release',
-                      value: ['1*'],
-                    },
-                  ],
-                },
-                id: 39,
-              },
               {
                 sampleRate: 0.2,
                 type: 'trace',
@@ -70,21 +57,6 @@ describe('Filters and Sampling - Transaction rule', function () {
         body: TestStubs.Project({
           dynamicSampling: {
             rules: [
-              {
-                sampleRate: 0.1,
-                type: 'error',
-                condition: {
-                  op: 'and',
-                  inner: [
-                    {
-                      op: 'glob',
-                      name: 'event.release',
-                      value: ['1*'],
-                    },
-                  ],
-                },
-                id: 44,
-              },
               {
                 sampleRate: 0.6,
                 type: 'trace',
@@ -114,24 +86,14 @@ describe('Filters and Sampling - Transaction rule', function () {
 
       renderComponent();
 
-      // Error rules container
-      expect(
-        screen.queryByText('There are no error rules to display')
-      ).not.toBeInTheDocument();
-      expect(screen.getByText('Errors only')).toBeInTheDocument();
-
       // Transaction traces and individual transactions rules container
       expect(
         screen.queryByText('There are no transaction rules to display')
       ).not.toBeInTheDocument();
-      const transactionTraceRules = screen.getByText('Transaction traces');
-      expect(transactionTraceRules).toBeInTheDocument();
-
-      const editRuleButtons = screen.getAllByLabelText('Edit Rule');
-      expect(editRuleButtons).toHaveLength(2);
+      expect(screen.getByText('Transaction traces')).toBeInTheDocument();
 
       // Open rule modal - edit transaction rule
-      await renderModal(editRuleButtons[1]);
+      await renderModal(screen.getByLabelText('Edit Rule'));
 
       // Modal content
       expect(screen.getByText('Edit Transaction Sampling Rule')).toBeInTheDocument();
@@ -142,28 +104,19 @@ describe('Filters and Sampling - Transaction rule', function () {
       expect(screen.getByLabelText('Search or add a release')).toBeInTheDocument();
 
       // Release field is not empty
-      const releaseFieldValues = screen.getByTestId('multivalue');
-      expect(releaseFieldValues).toBeInTheDocument();
-      expect(releaseFieldValues).toHaveTextContent('1.2.3');
+      expect(screen.getByTestId('multivalue')).toHaveTextContent('1.2.3');
 
       // Button is enabled - meaning the form is valid
-      const saveRuleButton = screen.getByLabelText('Save Rule');
-      expect(saveRuleButton).toBeInTheDocument();
-      expect(saveRuleButton).toBeEnabled();
-
-      // Sample rate field
-      const sampleRateField = screen.getByPlaceholderText('\u0025');
-      expect(sampleRateField).toBeInTheDocument();
+      expect(screen.getByLabelText('Save Rule')).toBeEnabled();
 
       // Sample rate is not empty
-      expect(sampleRateField).toHaveValue(20);
+      expect(screen.getByPlaceholderText('\u0025')).toHaveValue(20);
 
       // Clear release field
       userEvent.clear(screen.getByLabelText('Search or add a release'));
 
       // Release field is now empty
-      const newReleaseFieldValues = screen.queryByTestId('multivalue');
-      expect(newReleaseFieldValues).not.toBeInTheDocument();
+      expect(screen.queryByTestId('multivalue')).not.toBeInTheDocument();
 
       expect(screen.getByLabelText('Save Rule')).toBeDisabled();
 
@@ -171,40 +124,34 @@ describe('Filters and Sampling - Transaction rule', function () {
       userEvent.paste(screen.getByLabelText('Search or add a release'), '[0-9]');
 
       // Autocomplete suggests options
-      const autocompleteOption = screen.getByTestId('[0-9]');
-      expect(autocompleteOption).toBeInTheDocument();
-      expect(autocompleteOption).toHaveTextContent('[0-9]');
+      expect(screen.getByTestId('[0-9]')).toHaveTextContent('[0-9]');
 
       // Click on the suggested option
-      userEvent.click(autocompleteOption);
+      userEvent.click(screen.getByTestId('[0-9]'));
 
       expect(screen.getByLabelText('Save Rule')).toBeEnabled();
 
       // Clear sample rate field
-      userEvent.clear(sampleRateField);
+      userEvent.clear(screen.getByPlaceholderText('\u0025'));
 
       expect(screen.getByLabelText('Save Rule')).toBeDisabled();
 
       // Update sample rate field
-      userEvent.paste(sampleRateField, '60');
+      userEvent.paste(screen.getByPlaceholderText('\u0025'), '60');
 
       // Save button is now enabled
-      const saveRuleButtonEnabled = screen.getByLabelText('Save Rule');
-      expect(saveRuleButtonEnabled).toBeEnabled();
+      expect(screen.getByLabelText('Save Rule')).toBeEnabled();
 
       // Click on save button
-      userEvent.click(saveRuleButtonEnabled);
+      userEvent.click(screen.getByLabelText('Save Rule'));
 
       // Modal will close
       await waitForElementToBeRemoved(() =>
         screen.queryByText('Edit Transaction Sampling Rule')
       );
 
-      // Error rules panel is updated
-      expect(screen.getByText('Errors only')).toBeInTheDocument();
-
       expect(screen.getByText('Transaction traces')).toBeInTheDocument();
-      expect(screen.getAllByText('Release')).toHaveLength(2);
+      expect(screen.getByText('Release')).toBeInTheDocument();
 
       // Old values
       expect(screen.queryByText('1.2.3')).not.toBeInTheDocument();
@@ -259,9 +206,7 @@ describe('Filters and Sampling - Transaction rule', function () {
         expect(screen.getByText('Sampling Rate \u0025')).toBeInTheDocument();
         expect(screen.getByPlaceholderText('\u0025')).toHaveValue(null);
         expect(screen.getByLabelText('Cancel')).toBeInTheDocument();
-        const saveRuleButton = screen.getByLabelText('Save Rule');
-        expect(saveRuleButton).toBeInTheDocument();
-        expect(saveRuleButton).toBeDisabled();
+        expect(screen.getByLabelText('Save Rule')).toBeDisabled();
 
         // Close Modal
         userEvent.click(screen.getByLabelText('Close Modal'));
@@ -283,9 +228,10 @@ describe('Filters and Sampling - Transaction rule', function () {
         expect(screen.getByText(/filter conditions/i)).toBeInTheDocument();
 
         // Tracing Condition Options
-        const modal = screen.getByRole('dialog');
         conditionTracingCategories.forEach(conditionTracingCategory => {
-          expect(within(modal).getByText(conditionTracingCategory)).toBeInTheDocument();
+          expect(
+            within(screen.getByRole('dialog')).getByText(conditionTracingCategory)
+          ).toBeInTheDocument();
         });
 
         // Unchecked tracing checkbox
@@ -296,7 +242,9 @@ describe('Filters and Sampling - Transaction rule', function () {
 
         // No Tracing Condition Options
         commonConditionCategories.forEach(commonConditionCategory => {
-          expect(within(modal).getByText(commonConditionCategory)).toBeInTheDocument();
+          expect(
+            within(screen.getByRole('dialog')).getByText(commonConditionCategory)
+          ).toBeInTheDocument();
         });
 
         // Close Modal
@@ -362,36 +310,28 @@ describe('Filters and Sampling - Transaction rule', function () {
           expect(screen.getByLabelText('Search or add a release')).toBeInTheDocument();
 
           // Release field is empty
-          const releaseFieldValues = screen.queryByTestId('multivalue');
-          expect(releaseFieldValues).not.toBeInTheDocument();
+          expect(screen.queryByTestId('multivalue')).not.toBeInTheDocument();
 
           // Type into realease field
           userEvent.paste(screen.getByLabelText('Search or add a release'), '1.2.3');
 
           // Autocomplete suggests options
-          const autocompleteOption = screen.getByTestId('1.2.3');
-          expect(autocompleteOption).toBeInTheDocument();
-          expect(autocompleteOption).toHaveTextContent('1.2.3');
+          expect(screen.getByTestId('1.2.3')).toHaveTextContent('1.2.3');
 
           // Click on the suggested option
-          userEvent.click(autocompleteOption);
+          userEvent.click(screen.getByTestId('1.2.3'));
 
           // Button is still disabled
-          const saveRuleButton = screen.getByLabelText('Save Rule');
-          expect(saveRuleButton).toBeInTheDocument();
-          expect(saveRuleButton).toBeDisabled();
+          expect(screen.getByLabelText('Save Rule')).toBeDisabled();
 
           // Fill sample rate field
-          const sampleRateField = screen.getByPlaceholderText('\u0025');
-          expect(sampleRateField).toBeInTheDocument();
-          userEvent.paste(sampleRateField, '20');
+          userEvent.paste(screen.getByPlaceholderText('\u0025'), '20');
 
           // Save button is now enabled
-          const saveRuleButtonEnabled = screen.getByLabelText('Save Rule');
-          expect(saveRuleButtonEnabled).toBeEnabled();
+          expect(screen.getByLabelText('Save Rule')).toBeEnabled();
 
           // Click on save button
-          userEvent.click(saveRuleButtonEnabled);
+          userEvent.click(screen.getByLabelText('Save Rule'));
 
           // Modal will close
           await waitForElementToBeRemoved(() =>
@@ -402,8 +342,7 @@ describe('Filters and Sampling - Transaction rule', function () {
           expect(
             screen.queryByText('There are no transaction rules to display')
           ).not.toBeInTheDocument();
-          const transactionTraceRules = screen.getByText('Transaction traces');
-          expect(transactionTraceRules).toBeInTheDocument();
+          expect(screen.getByText('Transaction traces')).toBeInTheDocument();
           expect(screen.getByText('Release')).toBeInTheDocument();
           expect(screen.getByText('1.2.3')).toBeInTheDocument();
           expect(screen.getByText('20%')).toBeInTheDocument();
@@ -462,36 +401,30 @@ describe('Filters and Sampling - Transaction rule', function () {
             expect(screen.getByLabelText('Search or add a release')).toBeInTheDocument();
 
             // Release field is empty
-            const releaseFieldValues = screen.queryByTestId('multivalue');
-            expect(releaseFieldValues).not.toBeInTheDocument();
+            expect(screen.queryByTestId('multivalue')).not.toBeInTheDocument();
 
             // Type into realease field
             userEvent.paste(screen.getByLabelText('Search or add a release'), '1.2.3');
 
             // Autocomplete suggests options
-            const autocompleteOption = screen.getByTestId('1.2.3');
-            expect(autocompleteOption).toBeInTheDocument();
-            expect(autocompleteOption).toHaveTextContent('1.2.3');
+            expect(screen.getByTestId('1.2.3')).toHaveTextContent('1.2.3');
 
             // Click on the suggested option
-            userEvent.click(autocompleteOption);
+            userEvent.click(screen.getByTestId('1.2.3'));
 
             // Button is still disabled
-            const saveRuleButton = screen.getByLabelText('Save Rule');
-            expect(saveRuleButton).toBeInTheDocument();
-            expect(saveRuleButton).toBeDisabled();
+
+            expect(screen.getByLabelText('Save Rule')).toBeDisabled();
 
             // Fill sample rate field
-            const sampleRateField = screen.getByPlaceholderText('\u0025');
-            expect(sampleRateField).toBeInTheDocument();
-            userEvent.paste(sampleRateField, '20');
+
+            userEvent.paste(screen.getByPlaceholderText('\u0025'), '20');
 
             // Save button is now enabled
-            const saveRuleButtonEnabled = screen.getByLabelText('Save Rule');
-            expect(saveRuleButtonEnabled).toBeEnabled();
+            expect(screen.getByLabelText('Save Rule')).toBeEnabled();
 
             // Click on save button
-            userEvent.click(saveRuleButtonEnabled);
+            userEvent.click(screen.getByLabelText('Save Rule'));
 
             // Modal will close
             await waitForElementToBeRemoved(() =>
@@ -502,10 +435,7 @@ describe('Filters and Sampling - Transaction rule', function () {
             expect(
               screen.queryByText('There are no transaction rules to display')
             ).not.toBeInTheDocument();
-            const individualTransactionRules = screen.getByText(
-              'Individual transactions'
-            );
-            expect(individualTransactionRules).toBeInTheDocument();
+            expect(screen.getByText('Individual transactions')).toBeInTheDocument();
             expect(screen.getByText('Release')).toBeInTheDocument();
             expect(screen.getByText('1.2.3')).toBeInTheDocument();
             expect(screen.getByText('20%')).toBeInTheDocument();
@@ -552,16 +482,15 @@ describe('Filters and Sampling - Transaction rule', function () {
 
             // Open Modal
             await renderModal(screen.getByText('Add transaction rule'));
-            const checkedCheckbox = screen.getByRole('checkbox');
 
             // Checked tracing checkbox
-            expect(checkedCheckbox).toBeChecked();
+            expect(screen.getByRole('checkbox')).toBeChecked();
 
             // Uncheck tracing checkbox
-            userEvent.click(checkedCheckbox);
+            userEvent.click(screen.getByRole('checkbox'));
 
             // Unched tracing checkbox
-            expect(checkedCheckbox).not.toBeChecked();
+            expect(screen.getByRole('checkbox')).not.toBeChecked();
 
             // Click on 'Add condition'
             userEvent.click(screen.getByText('Add Condition'));
@@ -601,21 +530,16 @@ describe('Filters and Sampling - Transaction rule', function () {
             }
 
             // Button is still disabled
-            const saveRuleButton = screen.getByLabelText('Save Rule');
-            expect(saveRuleButton).toBeInTheDocument();
-            expect(saveRuleButton).toBeDisabled();
+            expect(screen.getByLabelText('Save Rule')).toBeDisabled();
 
             // Fill sample rate field
-            const sampleRateField = screen.getByPlaceholderText('\u0025');
-            expect(sampleRateField).toBeInTheDocument();
-            userEvent.paste(sampleRateField, '20');
+            userEvent.paste(screen.getByPlaceholderText('\u0025'), '20');
 
             // Save button is now enabled
-            const saveRuleButtonEnabled = screen.getByLabelText('Save Rule');
-            expect(saveRuleButtonEnabled).toBeEnabled();
+            expect(screen.getByLabelText('Save Rule')).toBeEnabled();
 
             // Click on save button
-            userEvent.click(saveRuleButtonEnabled);
+            userEvent.click(screen.getByLabelText('Save Rule'));
 
             // Modal will close
             await waitForElementToBeRemoved(() =>
@@ -626,10 +550,7 @@ describe('Filters and Sampling - Transaction rule', function () {
             expect(
               screen.queryByText('There are no transaction rules to display')
             ).not.toBeInTheDocument();
-            const individualTransactionRules = screen.getByText(
-              'Individual transactions'
-            );
-            expect(individualTransactionRules).toBeInTheDocument();
+            expect(screen.getByText('Individual transactions')).toBeInTheDocument();
             expect(screen.getByText('Legacy Browser')).toBeInTheDocument();
             for (const legacyBrowser of legacyBrowsers) {
               const {title} = LEGACY_BROWSER_LIST[legacyBrowser];
@@ -650,21 +571,6 @@ describe('Filters and Sampling - Transaction rule', function () {
         body: TestStubs.Project({
           dynamicSampling: {
             rules: [
-              {
-                sampleRate: 0.1,
-                type: 'error',
-                condition: {
-                  op: 'and',
-                  inner: [
-                    {
-                      op: 'glob',
-                      name: 'event.release',
-                      value: ['1*'],
-                    },
-                  ],
-                },
-                id: 39,
-              },
               {
                 sampleRate: 0.2,
                 type: 'transaction',
@@ -692,21 +598,6 @@ describe('Filters and Sampling - Transaction rule', function () {
         body: TestStubs.Project({
           dynamicSampling: {
             rules: [
-              {
-                sampleRate: 0.1,
-                type: 'error',
-                condition: {
-                  op: 'and',
-                  inner: [
-                    {
-                      op: 'glob',
-                      name: 'event.release',
-                      value: ['1*'],
-                    },
-                  ],
-                },
-                id: 44,
-              },
               {
                 sampleRate: 0.6,
                 type: 'transaction',
@@ -736,24 +627,14 @@ describe('Filters and Sampling - Transaction rule', function () {
 
       renderComponent();
 
-      // Error rules container
-      expect(
-        screen.queryByText('There are no error rules to display')
-      ).not.toBeInTheDocument();
-      expect(screen.getByText('Errors only')).toBeInTheDocument();
-
       // Transaction traces and individual transactions rules container
       expect(
         screen.queryByText('There are no transaction rules to display')
       ).not.toBeInTheDocument();
-      const transactionTraceRules = screen.getByText('Individual transactions');
-      expect(transactionTraceRules).toBeInTheDocument();
-
-      const editRuleButtons = screen.getAllByLabelText('Edit Rule');
-      expect(editRuleButtons).toHaveLength(2);
+      expect(screen.getByText('Individual transactions')).toBeInTheDocument();
 
       // Open rule modal - edit transaction rule
-      await renderModal(editRuleButtons[1]);
+      await renderModal(screen.getByLabelText('Edit Rule'));
 
       // Modal content
       expect(screen.getByText('Edit Transaction Sampling Rule')).toBeInTheDocument();
@@ -767,23 +648,16 @@ describe('Filters and Sampling - Transaction rule', function () {
       expect(screen.getByTestId('multivalue')).toBeInTheDocument();
 
       // Button is enabled - meaning the form is valid
-      const saveRuleButton = screen.getByLabelText('Save Rule');
-      expect(saveRuleButton).toBeInTheDocument();
-      expect(saveRuleButton).toBeEnabled();
-
-      // Sample rate field
-      const sampleRateField = screen.getByPlaceholderText('\u0025');
-      expect(sampleRateField).toBeInTheDocument();
+      expect(screen.getByLabelText('Save Rule')).toBeEnabled();
 
       // Sample rate is not empty
-      expect(sampleRateField).toHaveValue(20);
+      expect(screen.getByPlaceholderText('\u0025')).toHaveValue(20);
 
       // Clear release field
       userEvent.clear(screen.getByLabelText('Search or add a release'));
 
       // Release field is now empty
-      const newReleaseFieldValues = screen.queryByTestId('multivalue');
-      expect(newReleaseFieldValues).not.toBeInTheDocument();
+      expect(screen.queryByTestId('multivalue')).not.toBeInTheDocument();
 
       expect(screen.getByLabelText('Save Rule')).toBeDisabled();
 
@@ -791,40 +665,34 @@ describe('Filters and Sampling - Transaction rule', function () {
       userEvent.paste(screen.getByLabelText('Search or add a release'), '[0-9]');
 
       // Autocomplete suggests options
-      const autocompleteOption = screen.getByTestId('[0-9]');
-      expect(autocompleteOption).toBeInTheDocument();
-      expect(autocompleteOption).toHaveTextContent('[0-9]');
+      expect(screen.getByTestId('[0-9]')).toHaveTextContent('[0-9]');
 
       // Click on the suggested option
-      userEvent.click(autocompleteOption);
+      userEvent.click(screen.getByTestId('[0-9]'));
 
       expect(screen.getByLabelText('Save Rule')).toBeEnabled();
 
       // Clear sample rate field
-      userEvent.clear(sampleRateField);
+      userEvent.clear(screen.getByPlaceholderText('\u0025'));
 
       expect(screen.getByLabelText('Save Rule')).toBeDisabled();
 
       // Update sample rate field
-      userEvent.paste(sampleRateField, '60');
+      userEvent.paste(screen.getByPlaceholderText('\u0025'), '60');
 
       // Save button is now enabled
-      const saveRuleButtonEnabled = screen.getByLabelText('Save Rule');
-      expect(saveRuleButtonEnabled).toBeEnabled();
+      expect(screen.getByLabelText('Save Rule')).toBeEnabled();
 
       // Click on save button
-      userEvent.click(saveRuleButtonEnabled);
+      userEvent.click(screen.getByLabelText('Save Rule'));
 
       // Modal will close
       await waitForElementToBeRemoved(() =>
         screen.queryByText('Edit Transaction Sampling Rule')
       );
 
-      // Error rules panel is updated
-      expect(screen.getByText('Errors only')).toBeInTheDocument();
-
       expect(screen.getByText('Individual transactions')).toBeInTheDocument();
-      expect(screen.getAllByText('Release')).toHaveLength(2);
+      expect(screen.getByText('Release')).toBeInTheDocument();
 
       // Old values
       expect(screen.queryByText('1.2.3')).not.toBeInTheDocument();
